feat(GA): moved the structs to own .h
