+++ apps/preference-solver/src/Pref_Solver/pref.h
@@ -1,41 +1,22 @@
 #include <string>
 #include <iostream>
 #include <vector>
+#include "types.h"
+#include <unordered_map>
+
 using std::cout;
 using std::endl;
 using std::string;
 
 class GA_PREF
 {
-};
-enum Day
-{
-    MONDAY,
-    TUEDAY,
-    WEDNESDAY,
-    THURSDAY,
-    FRIDAY,
-    SATURDAY,
-    SUNDAY
-};
-
-struct Event
-{
-    Day eventDay;
-    string eventID ;
-    string moduleCode;
-    int event_start;
-    int event_end;
-};
-
-struct Module
-{
-    string moduleCode;
-    int number_Occur;
-    std::vector<Event> events;
-};
-struct GA_Data
-{
-    std::vector<Module> modules;
+private:
     int targetTime;
+    std::unordered_map<string, int> numOccurences; 
+    // finds how many times module code should occcur
+
+public:
+    GA_PREF(GA_Data data)
+    {
+    }
 };
