+++ apps/preference-solver/src/Pref_Solver/types.h
@@ -1,0 +1,45 @@
+#include <string>
+#include <iostream>
+#include <vector>
+using std::cout;
+using std::endl;
+using std::string;
+
+enum Day
+{
+    MONDAY,
+    TUEDAY,
+    WEDNESDAY,
+    THURSDAY,
+    FRIDAY,
+    SATURDAY,
+    SUNDAY
+};
+
+// Listed with GA since there may be a direct mapping from json required later to a struct
+struct EventGA
+{
+    Day eventDay;
+    string eventID;
+    string moduleCode;
+    int event_start;
+    int event_end;
+    bool is_active;
+};
+
+struct ModuleGA
+{
+    string moduleCode;
+    int number_Occur;
+    std::vector<EventGA> events;
+};
+
+struct GA_Data
+{
+    std::vector<ModuleGA> modules;
+    int targetTime;
+};
+
+struct EventChromosome
+{
+};
